feat!: move generation of grammar files to an `init` command
The generate subcommand should stick to solely generating a parser and
headers.
