+++ cli/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod fuzz;
 pub mod generate;
 pub mod highlight;
+pub mod init;
 pub mod logger;
 pub mod parse;
 pub mod playground;
