+++ cli/src/templates/.editorconfig
