+++ cli/src/templates/PARSER_NAME.pc.in
