+++ cli/src/templates/__init__.py
