+++ cli/src/templates/__init__.pyi
