+++ cli/src/templates/_cargo.toml
