+++ cli/src/templates/binding.go
