+++ cli/src/templates/binding.gyp
