+++ cli/src/templates/binding_test.go
