+++ cli/src/templates/binding_test.js
