+++ cli/src/templates/gitattributes
