+++ cli/src/templates/gitignore
