+++ cli/src/templates/go.mod
