+++ cli/src/templates/index.d.ts
