+++ cli/src/templates/index.js
