+++ cli/src/templates/js-binding.cc
