+++ cli/src/templates/lib.rs
