+++ cli/src/templates/package.json
