+++ cli/src/templates/package.swift
