+++ cli/src/templates/py-binding.c
