+++ cli/src/templates/pyproject.toml
