+++ cli/src/templates/setup.py
