+++ cli/src/templates/test_binding.py
