+++ cli/src/templates/tests.swift
